Add styling to info modal to make look consistent
The look of the info modal used to change from backend to frontend
because the base font-size is different on each end. This standardizes
the modal to have the same look for consistency. It also changes the text
in the button from "?" to "User Guide" as that's a bit more clear.