Seperate frontend users to avoid sharing carts
This removes the ID specification from the fake user generation. It was there
because one URL on the backend expects an ID from try_spree_current_user to take
you to the user edit page. Since the user logged in doesn't actually exist,
I've changed this URL to just take you to the first user in the database
(which is the original admin user)