Allow checkout with new saving schema
Adds a checkout decorator that always marks the address as temporary -
which disables addressbook saving. This is important because it tries
to save the address under the fake user we're passing to foil authentication.

Also adds a billing_name method - the sample frontend is relying on
changes from an unreleased Solidus version, so this is a temporary fix
until Solidus 2.11 comes out.