Allow changes to sample data
Creates a SampleChanges column that records any changes made to sample object.
If the object is a sample object, it'll record the changes in that column
instead of adjusting the original object. When it finds an object that
has an associated SampleChange, it'll set the attributes for that object
to the SampleChange data instead, allowing the user that made the change
to see their own changes - but because SampleChange is scoped to the user,
no other user will see those changes. :tada: