Add rake task to destroy old sample data
Old sample data is bogging the database down. Ideally, we don't want
to destroy records while people are still using the demo, so we'll
only destroy data that hasn't been updated in weeks. Initially,
this only handles orders and products - the bulk of the excess records.
But in the future we may need to expand this to other records, as I
imagine the lesser used records will pile up over time (and really,
our available usage is only like ~1000-3000 records)