Fix stock item locking with unpersisted changes
Currently there's a bug where if you check out using the same sample
product variant twice, you'll trigger a runtime error. This is because
when you "edit" sample data, what you're actually doing is creating
a SampleChange where that change is stored. The next time you fetch
the thing you tried to change, the SampleChange you made will be loaded
in with it, and whatever you changed will be swapped in.

This swap isn't persisted though, and so when StockItem has `with_lock`
called on it and there's already a SampleChange associated with it,
the unpersisted data causes this error.

Since we're not actually saving anything on the stock_item, I've
disabled with_lock for changes as long as it's sample data.
Everything else will be locked as normal!