Make solidus-demo compatible with the GraphQL extension
It makes this app compatible with the GraphQL extension that was added
in #89. However, as with the HTML version, it adds compatibility only for
guest checkouts, which are taken as coming from a `test@solidus.io`
user. Authenticated checkouts won't work because the singleton user is
not persisted, so the order can't be recorded along with the user.

We needed to monkey patch the GraphQL extension, but it was tricky to
support the `prependers` gem setup for Rails + external files overrides. As
a consequence, we've taken the opportunity to remove the dependency on that
library, and instead we're performing regular monkey patching as it's now
[suggested in the rails
guides](https://guides.rubyonrails.org/engines.html#improving-engine-functionality).

We're also tweaking some tests to avoid flaky tests because of shared
global state.