Fix cart link bubble
Expected behavior
-----------------

Given the cart has items
When I view the cart
Then I should see that it has a bubble with the number of items on it

Actual behavior
---------------

The cart link shows the text "CART: (<count>)" on top of the cart icon.
It also shows the cart total amount.

Cause
-----

In `app/monkey_patches/spree/base_helper_monkey_patch.rb`, it seems that
prepending the module with the modified `link_to_cart` helper to
`Spree::BaseHelper` does not work as expected.